requirements.txt
aniso8601==9.0.1
asyncio==3.4.3
attrs==22.2.0
Authlib==1.2.0
bcrypt==4.0.1
Bootstrap-Flask==2.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Levenshtein==0.20.9
MarkupSafe==2.1.1
marshmallow==3.19.0
mysql-connector-python==8.0.31
packaging==22.0
pluggy==1.0.0
protobuf==3.20.1
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.2
pytest-flask==1.2.0
python-Levenshtein==0.20.9
pytz==2022.6
rapidfuzz==2.13.7
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.46
thefuzz==0.19.0
urllib3==1.26.13
Werkzeug==2.2.2
WTForms==3.0.1
